pyproject.toml
[tool.poetry]
name = "api-gateway"
version = "0.1.0"
package-mode = false
description = ""
authors = ["ndestrieux <n.destrieux@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.12"
fastapi = {extras = ["standard"], version = "~0.115"}
pydantic-settings = "~2.2"
pyjwt = {extras = ["crypto"], version = "~2.8"}
uvicorn = {extras = ["standard"], version = "~0.29"}
httpx = "~0.28"


[tool.poetry.group.dev.dependencies]
black = "*"
isort = "*"
pre-commit = "~3.7"
pytest = "~8.3"
pytest-asyncio = "~0.23"
pytest-env = "~1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests"]
asyncio_mode = "strict"
addopts = "-v -ra -q"
minversion = "6.0"
env = [
    "auth0_domain=test_domain",
    "auth0_api_audience=test_audience",
    "auth0_issuer=test_issuer",
    "auth0_algorithms=test_algorithms",
    "library_endpoint=test_library_endpoint",
    "forum_endpoint=test_forum_endpoint"
]
log_cli = true
log_cli_level = "WARNING"
log_cli_format = "%(asctime)s %(levelname)s %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
